First version of modified flaperon-mixer for aileron differential
The flaperon-mixer UPUP is used to realize a APM-controlled aileron
differential for big soaring airplanes to reduce drag and slip. The
positive value of the aileron is reduced by a scale-factor which is
hard-coded at this point. The diff-value is 70%.
To use the aileron-diff, set RC_5_function and RC_6_function to
flaperon setting (24+25). The flaperon-mixer ist DISABLED at this
release, because the code needs to be tested. The magic is done in
function Plane::channel_output_mixer(). Maybe there could be a better
place, to realize the differential in future.